requirements.txt
fastapi
uvicorn
httpx
sqlmodel
python-dotenv
# psycopg2-binary
psycopg2